requirements.txt
praw
google-api-python-client
djangorestframework
django  